Don't remove non-existent extract dir

+++ src/Microsoft.DotNet.SignTool/src/ZipData.cs
@@ -300,9 +300,11 @@ namespace Microsoft.DotNet.SignTool
             }
             finally
             {
-                Directory.Delete(extractDir, recursive: true);
+                if (Directory.Exists(extractDir))
+                {
+                    Directory.Delete(extractDir, recursive: true);
+                }
             }
-        }
 
         private void RepackPkgOrAppBundles(TaskLoggingHelper log, string tempDir, string pkgToolPath)
         {
